+++ CHANGELOG.md
@@ -46,6 +46,10 @@ The type system was made possible thanks to a partnership between [CNRS](https:/
   * [Kernel] Warn on unused requires
   * [Regex] Add `Regex.import/1` to import regexes defined with `/E`
 
+#### ExUnit
+
+  * [ExUnit.CaptureLog] Add `:formatter` option for custom log formatting
+
 #### Mix
 
   * [mix test] Add `mix test --dry-run`